+++ autoload/_globals.py
@@ -1,0 +1,6 @@
+from enum import Enum
+
+
+class LoadType(Enum):
+    func = "function"
+    clazz = "class"
